+++ docs/CHANGELOG.md
@@ -12,10 +12,11 @@ Alle viktige endringer i NISSY-scriptene vil bli dokumentert i denne filen.
 ---
 ## [1.2.0] - 04.01.2026
 
-### 🚕 Ressursinfo forbedret!
+### 🎉 Utvidelse med nytt script Bestillingsmodul.js og 🚕 Ressursinfo forbedret!
 
 #### Hva er nytt?
 - **Ressursinfo.js** - Parser 2000 XML og henter ut navn, adresse, planlagt tid med korrekt æøå-formatering. Forbedret UI-visning.
+- **Bestillingsmodul.js** - Lar deg velge foretrukken modul mellom 4-stegs og ensides og husker valget for sesjonen. Ligger i alle script-pakkene.
 
 ## [1.1.0] - 02.01.2026